Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/report-uri-in-meta-tag-ignored.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="report-uri /security/contentSecurityPolicy/resources/save-report.php">
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
<p>This tests that a console message is logged to explain that the Content Security Policy <code>report-uri</code> directive is ignored when contained in a policy defined via an HTML meta element.</p>
</body>
</html>
